+++ packages/host-cache/src/components/index.js
@@ -1,0 +1,3 @@
+// @flow
+import Host from "./host";
+export { Host };
